Refactor code to implement SendGrid API within action logic for adding user to DB
