+++ awwficial-rentals-website/next.config.js
@@ -2,7 +2,7 @@
 const nextConfig = {
     experimental: {
         serverActions: true
-    }
+    },
 }
 
 module.exports = nextConfig
